refactor : 유틸 함수 위치 src/util로 수정
